BJS2-77902: hotfix bug in AchievementCache with serialization and delete double RedisConfig.java
