--- src/main/java/faang/school/achievement/config/context/redis/RedisConfig.java
@@ -1,93 +1,0 @@
-package faang.school.achievement.config.context.redis;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import faang.school.achievement.listener.CommentEventListener;
-import faang.school.achievement.model.Achievement;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.CacheManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheConfiguration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.RedisSerializationContext;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-import java.time.Duration;
-
-@Configuration
-public class RedisConfig {
-
-    @Value("${cache.redis.ttl}")
-    private long ttl;
-
-    @Value("${spring.data.redis.channel.comment}")
-    private String commentChannel;
-
-    @Bean
-    public RedisCacheConfiguration redisCacheConfiguration() {
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.registerModule(new JavaTimeModule());
-        mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-        Jackson2JsonRedisSerializer<Achievement> serializer = new Jackson2JsonRedisSerializer<>(mapper, Achievement.class);
-
-        return RedisCacheConfiguration.defaultCacheConfig()
-                .entryTtl(Duration.ofSeconds(ttl))
-                .serializeValuesWith(
-                        RedisSerializationContext.SerializationPair.fromSerializer(serializer)
-                )
-                .disableCachingNullValues();
-    }
-
-    @Bean
-    public CacheManager redisCacheManager(
-            RedisConnectionFactory redisConnectionFactory,
-            RedisCacheConfiguration redisCacheConfiguration) {
-        return RedisCacheManager.builder(redisConnectionFactory)
-                .cacheDefaults(redisCacheConfiguration)
-                .build();
-    }
-
-    @Bean
-    public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory connectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(connectionFactory);
-
-        Jackson2JsonRedisSerializer<Object> serializer = new Jackson2JsonRedisSerializer<>(Object.class);
-        template.setDefaultSerializer(serializer);
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(serializer);
-        return template;
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisContainer(
-            RedisConnectionFactory connectionFactory,
-            @Qualifier("listenerCommentEventAdapter") MessageListenerAdapter listenerCommentEvetAdapter,
-            @Qualifier("commentTopic") ChannelTopic commentTopic) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(connectionFactory);
-        container.addMessageListener(listenerCommentEvetAdapter, commentTopic);
-        return container;
-    }
-
-    @Bean
-    @Qualifier("listenerCommentEventAdapter")
-    public MessageListenerAdapter listenerCommentEvetAdapter(CommentEventListener commentEventListener) {
-        return new MessageListenerAdapter(commentEventListener);
-    }
-
-    @Bean
-    @Qualifier("commentTopic")
-    public ChannelTopic commentTopic() {
-        return new ChannelTopic(commentChannel);
-    }
-}
