+++ src/main/java/faang/school/achievement/config/redis/RedisConfig.java
@@ -1,22 +1,82 @@
 package faang.school.achievement.config.redis;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import faang.school.achievement.dto.cache.AchievementCacheDto;
+import faang.school.achievement.listener.AchievementEventListener;
+import faang.school.achievement.listener.CommentEventListener;
+import faang.school.achievement.model.Achievement;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cache.CacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.cache.RedisCacheConfiguration;
+import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
+import org.springframework.data.redis.serializer.RedisSerializationContext;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
-import faang.school.achievement.listener.AchievementEventListener;
-import org.springframework.beans.factory.annotation.Value;
+import java.time.Duration;
 
 @Configuration
 public class RedisConfig {
+
+    @Value("${cache.redis.ttl}")
+    private long ttl;
+
+    @Value("${spring.data.redis.channels.comment}")
+    private String commentChannel;
+
     @Value("${spring.data.redis.channels.achievement}")
     private String achievementChannel;
+
+    @Bean
+    public RedisCacheConfiguration redisCacheConfiguration() {
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new JavaTimeModule());
+        mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+        Jackson2JsonRedisSerializer<Achievement> serializer = new Jackson2JsonRedisSerializer<>(mapper, Achievement.class);
+
+        return RedisCacheConfiguration.defaultCacheConfig()
+                .entryTtl(Duration.ofSeconds(ttl))
+                .serializeValuesWith(
+                        RedisSerializationContext.SerializationPair.fromSerializer(serializer)
+                )
+                .disableCachingNullValues();
+    }
+
+    @Bean
+    public CacheManager redisCacheManager(
+            RedisConnectionFactory redisConnectionFactory,
+            RedisCacheConfiguration redisCacheConfiguration) {
+        return RedisCacheManager.builder(redisConnectionFactory)
+                .cacheDefaults(redisCacheConfiguration)
+                .build();
+    }
+
+    @Bean
+    public RedisTemplate<String, AchievementCacheDto> achievementRedisTemplate(
+            RedisConnectionFactory connectionFactory,
+            ObjectMapper objectMapper
+    ) {
+        RedisTemplate<String, AchievementCacheDto> template = new RedisTemplate<>();
+        template.setConnectionFactory(connectionFactory);
+
+        Jackson2JsonRedisSerializer<AchievementCacheDto> serializer =
+                new Jackson2JsonRedisSerializer<>(objectMapper, AchievementCacheDto.class);
+
+        template.setKeySerializer(new StringRedisSerializer());
+        template.setValueSerializer(serializer);
+        template.setDefaultSerializer(serializer);
+
+        return template;
+    }
 
     @Bean
     public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory connectionFactory) {
@@ -33,12 +93,25 @@ public class RedisConfig {
     @Bean
     public RedisMessageListenerContainer redisContainer(
             RedisConnectionFactory connectionFactory,
+            MessageListenerAdapter listenerCommentEvetAdapter,
             MessageListenerAdapter listenerAnalyticsEventAdapter,
+            ChannelTopic commentTopic,
             ChannelTopic analyticsTopic) {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
         container.setConnectionFactory(connectionFactory);
+        container.addMessageListener(listenerCommentEvetAdapter, commentTopic);
         container.addMessageListener(listenerAnalyticsEventAdapter, analyticsTopic);
         return container;
+    }
+
+    @Bean
+    public MessageListenerAdapter listenerCommentEvetAdapter(CommentEventListener commentEventListener) {
+        return new MessageListenerAdapter(commentEventListener);
+    }
+
+    @Bean
+    public ChannelTopic commentTopic() {
+        return new ChannelTopic(commentChannel);
     }
 
     @Bean
