+++ src/main/java/faang/school/achievement/service/AchievementService.java
@@ -1,6 +1,6 @@
 package faang.school.achievement.service;
 
-import faang.school.achievement.cash.AchievementCache;
+import faang.school.achievement.cache.AchievementCache;
 import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.AchievementProgress;
 import faang.school.achievement.repository.AchievementProgressRepository;
