+++ src/main/resources/application.yaml
@@ -21,7 +21,7 @@ spring:
     redis:
       port: 6379
       host: localhost
-      channel:
+      channels:
         achievement: achievement_channel
         follower: follower_channel
         comment: comment_channel
